Completed: Get & Create for notes
