+++ src/components/Main.jsx
@@ -9,10 +9,10 @@ import AboutPage from "../pages/AboutPage";
 function Main(props) {
   const [notes, setNotes] = useState(null);
 
-  const URL = "https://vibe-voyage.herokuapp.com/notes";
+  const URL = "https://vibe-voyage.herokuapp.com/";
 
   const getNotes = async () => {
-    const response = await fetch(URL);
+    const response = await fetch(URL + "notes/");
     const data = await response.json();
     console.log(data)
     setNotes(data);
@@ -20,7 +20,7 @@ function Main(props) {
 
   const createNotes = async (notes) => {
 
-    await fetch(URL, {
+    await fetch(URL + "notes/", {
       method: "POST",
       headers: {
         "Content-Type": "Application/json",
@@ -31,26 +31,36 @@ function Main(props) {
     getNotes();
   };
 
-const updateNotes = async (person, id) => {
-  await fetch(URL + id, {
-    method: "PUT",
-    headers: {
-      "Content-Type": "Application/json",
-    },
-    body: JSON.stringify(notes),
-  })
-  getNotes();
-}
+  const updateNotes = async (person, id) => {
+    await fetch(URL + "notes/" + id, {
+      method: "PUT",
+      headers: {
+        "Content-Type": "Application/json",
+      },
+      body: JSON.stringify(notes),
+    })
+    getNotes();
+  }
 
-const deleteNotes = async (id) => {
-  await fetch(URL + id,{
-    method:"DELETE",
-  }) 
-  getNotes();
-}
+  const deleteNotes = async (id) => {
+    await fetch(URL + "notes/" + id, {
+      method: "DELETE",
+    })
+    getNotes();
+  }
 
- 
-  useEffect(() => { getNotes() }, []);
+  const getVideos = async () => {
+    const response = await fetch(URL + "videos/");
+    const data = await response.json();
+    console.log(data)
+    setNotes(data);
+  };
+
+
+  useEffect(() => {
+    getNotes();
+
+  }, []);
 
 
   return (
